+++ src/datasources/asset/constants/constants.ts
@@ -1,0 +1,8 @@
+export enum AllFieldNames {
+    LOCATION = 'Location',
+    WORKSPACE = 'Workspace',
+    MODEL_NAME = 'ModelName',
+    VENDOR_NAME = 'VendorName',
+    BUS_TYPE = 'BusType',
+    ASSET_TYPE = 'AssetType',
+}
